requirements.txt
asgiref==3.7.2
asyncio==3.4.3
certifi==2023.7.22
charset-normalizer==3.3.2
django-cors-headers==4.3.1
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
urllib3==2.0.7
# To Serve Docker 
# gunicorn==20.1.0
